rename stats -> metrics
