+++ tokio/Cargo.toml
@@ -46,7 +46,6 @@ io-util = ["memchr", "bytes"]
 # stdin, stdout, stderr
 io-std = []
 macros = ["tokio-macros"]
-stats = []
 net = [
   "libc",
   "mio/os-poll",
@@ -84,6 +83,11 @@ signal = [
 sync = []
 test-util = ["rt", "sync", "time"]
 time = []
+
+# Technically, removing this is a breaking change even though it only ever did
+# anything with the unstable flag on. It is probably safe to get rid of it after
+# a few releases.
+stats = []
 
 [dependencies]
 tokio-macros = { version = "1.7.0", path = "../tokio-macros", optional = true }
