+++ tokio/src/runtime/metrics/mock.rs
@@ -1,8 +1,12 @@
-//! This file contains mocks of the types in src/runtime/stats/stats.rs
+//! This file contains mocks of the types in src/runtime/metrics
 
-pub(crate) struct RuntimeStats {}
+pub(crate) struct RuntimeMetrics {}
 
-impl RuntimeStats {
+pub(crate) struct WorkerMetrics {}
+
+pub(crate) struct MetricsBatch {}
+
+impl RuntimeMetrics {
     pub(crate) fn new(_worker_threads: usize) -> Self {
         Self {}
     }
@@ -10,25 +14,21 @@ impl RuntimeStats {
     /// Increment the number of tasks scheduled externally
     pub(crate) fn inc_remote_schedule_count(&self) {}
 
-    pub(crate) fn worker(&self, _index: usize) -> &WorkerStats {
-        &WorkerStats {}
+    pub(crate) fn worker(&self, _index: usize) -> &WorkerMetrics {
+        &WorkerMetrics {}
     }
 }
 
-pub(crate) struct WorkerStats {}
-
-pub(crate) struct WorkerStatsBatcher {}
-
-impl WorkerStats {
+impl WorkerMetrics {
     pub(crate) fn incr_stolen_count(&self, _n: u16) {}
 }
 
-impl WorkerStatsBatcher {
+impl MetricsBatch {
     pub(crate) fn new(_my_index: usize) -> Self {
         Self {}
     }
 
-    pub(crate) fn submit(&mut self, _to: &RuntimeStats) {}
+    pub(crate) fn submit(&mut self, _to: &RuntimeMetrics) {}
     pub(crate) fn about_to_park(&mut self) {}
     pub(crate) fn returned_from_park(&mut self) {}
     pub(crate) fn incr_poll_count(&mut self) {}
@@ -36,7 +36,7 @@ impl WorkerStatsBatcher {
 }
 
 cfg_rt_multi_thread! {
-    impl WorkerStatsBatcher {
+    impl MetricsBatch {
         pub(crate) fn incr_steal_count(&mut self, _by: u16) {}
         pub(crate) fn incr_overflow_count(&mut self, _by: u16) {}
     }
