+++ tokio/src/runtime/queue.rs
@@ -3,7 +3,7 @@
 use crate::loom::cell::UnsafeCell;
 use crate::loom::sync::atomic::{AtomicU16, AtomicU32};
 use crate::loom::sync::Arc;
-use crate::runtime::stats::{WorkerStats, WorkerStatsBatcher};
+use crate::runtime::{WorkerMetrics, MetricsBatch};
 use crate::runtime::task::{self, Inject};
 
 use std::mem::MaybeUninit;
@@ -106,7 +106,7 @@ impl<T> Local<T> {
         &mut self,
         mut task: task::Notified<T>,
         inject: &Inject<T>,
-        stats: &mut WorkerStatsBatcher,
+        metrics: &mut MetricsBatch,
     ) {
         let tail = loop {
             let head = self.inner.head.load(Acquire);
@@ -122,12 +122,12 @@ impl<T> Local<T> {
                 // Concurrently stealing, this will free up capacity, so only
                 // push the task onto the inject queue
                 inject.push(task);
-                stats.incr_overflow_count(1);
+                metrics.incr_overflow_count(1);
                 return;
             } else {
                 // Push the current task and half of the queue into the
                 // inject queue.
-                match self.push_overflow(task, real, tail, inject, stats) {
+                match self.push_overflow(task, real, tail, inject, metrics) {
                     Ok(_) => return,
                     // Lost the race, try again
                     Err(v) => {
@@ -169,7 +169,7 @@ impl<T> Local<T> {
         head: u16,
         tail: u16,
         inject: &Inject<T>,
-        stats: &mut WorkerStatsBatcher,
+        metrics: &mut MetricsBatch,
     ) -> Result<(), task::Notified<T>> {
         /// How many elements are we taking from the local queue.
         ///
@@ -254,7 +254,7 @@ impl<T> Local<T> {
         inject.push_batch(batch_iter.chain(std::iter::once(task)));
 
         // Add 1 to factor in the task currently being scheduled.
-        stats.incr_overflow_count(NUM_TASKS_TAKEN + 1);
+        metrics.incr_overflow_count(NUM_TASKS_TAKEN + 1);
 
         Ok(())
     }
@@ -310,8 +310,8 @@ impl<T> Steal<T> {
     pub(super) fn steal_into(
         &self,
         dst: &mut Local<T>,
-        dst_stats: &mut WorkerStatsBatcher,
-        src_stats: &WorkerStats,
+        dst_metrics: &mut MetricsBatch,
+        src_metrics: &WorkerMetrics,
     ) -> Option<task::Notified<T>> {
         // Safety: the caller is the only thread that mutates `dst.tail` and
         // holds a mutable reference.
@@ -337,8 +337,8 @@ impl<T> Steal<T> {
             return None;
         }
 
-        dst_stats.incr_steal_count(n);
-        src_stats.incr_stolen_count(n);
+        dst_metrics.incr_steal_count(n);
+        src_metrics.incr_stolen_count(n);
 
         // We are returning a task here
         n -= 1;
