+++ tokio/src/runtime/tests/queue.rs
@@ -1,5 +1,5 @@
 use crate::runtime::queue;
-use crate::runtime::stats::WorkerStatsBatcher;
+use crate::runtime::MetricsBatch;
 use crate::runtime::task::{self, Inject, Schedule, Task};
 
 use std::thread;
@@ -45,7 +45,7 @@ fn overflow() {
 
 #[test]
 fn steal_batch() {
-    let mut stats = WorkerStatsBatcher::new(0);
+    let mut metrics = MetricsBatch::new(0);
 
     let (steal1, mut local1) = queue::local();
     let (_, mut local2) = queue::local();
@@ -56,7 +56,7 @@ fn steal_batch() {
         local1.push_back(task, &inject);
     }
 
-    assert!(steal1.steal_into(&mut local2, &mut stats).is_some());
+    assert!(steal1.steal_into(&mut local2, &mut metrics).is_some());
 
     for _ in 0..1 {
         assert!(local2.pop().is_some());
