+++ tokio/src/runtime/thread_pool/mod.rs
@@ -9,9 +9,8 @@ pub(crate) use worker::Launch;
 pub(crate) use worker::block_in_place;
 
 use crate::loom::sync::Arc;
-use crate::runtime::stats::RuntimeStats;
 use crate::runtime::task::JoinHandle;
-use crate::runtime::{Callback, Parker};
+use crate::runtime::{Callback, Parker, RuntimeMetrics};
 
 use std::fmt;
 use std::future::Future;
@@ -103,8 +102,8 @@ impl Spawner {
         self.shared.close();
     }
 
-    pub(crate) fn stats(&self) -> &RuntimeStats {
-        self.shared.stats()
+    pub(crate) fn metrics(&self) -> &RuntimeMetrics {
+        self.shared.metrics()
     }
 }
